Add files via upload
This project is part of a course on Multivariate Analysis, focusing on clustering analysis to uncover patterns and groupings within multivariate data. Clustering analysis is a key technique in multivariate analysis, aimed at identifying natural groupings in datasets with multiple variables, thus revealing insights that are not immediately apparent through univariate or bivariate analysis.