Add files via upload
The objective of this project, as part of the "Database for Data Science" course, is to design and implement a Flood Relief Request Form for Jakarta's 2020 flood disaster. The primary goal of this form is to systematically collect and coordinate requests for goods and products needed by individuals affected by the flood in Jakarta in 2020, and to relay these requests to the DKI Jakarta government for appropriate action.