Add files via upload
This project is part of the "Data Mining and Business Intelligence" course, focusing on the implementation and evaluation of classification models, specifically using Support Vector Machines (SVM) and other classification techniques. The primary objective is to apply these models to a dataset, evaluate their performance, and extract meaningful business insights that can guide decision-making processes.