Add files via upload
This project is part of the "Statistical Computing" course, focusing on data wrangling techniques essential for preparing and transforming raw data into a usable format for statistical analysis. The primary objective is to equip students with practical skills in handling real-world data, ensuring it is clean, consistent, and ready for sophisticated statistical computations and analyses.