Add files via upload
This project, part of the "Econometrics" course, aims to investigate the effects of the Human Development Index (HDI) and Gross Regional Domestic Product (GRDP) on poverty levels in the regencies and cities of Central Java Province from 2011 to 2015. Using panel data analysis, the project seeks to uncover the dynamic relationship between these economic indicators and poverty, providing insights that can inform policy decisions to alleviate poverty in the region.