Add files via upload
This project for the "Unstructured Data Analysis" course focuses on performing sentiment analysis on Amazon product reviews. The objective is to classify the sentiments expressed in the reviews as positive, negative, or neutral, using both a lexicon-based method and the Support Vector Machine (SVM) algorithm. The project aims to compare the effectiveness of these approaches in accurately determining customer sentiment and providing actionable insights for businesses.