Add files via upload
This project for the "Digital Data Collection Methods" course focuses on the practical application of website scraping to collect and analyze data from the web. The objective is to equip students with the skills necessary to extract data from websites, process it, and extract meaningful insights. This project emphasizes best practices, ethical considerations, and the technical aspects of web scraping.